Add slugification for video and text chunk
Thought about several different options and landed on this:

`[type-of-chunk]-slug-chunkID`

For example:
A video chunk for a video entitled What Is FRED? | Federal Reserve Economic Database that was given a DB ID of 14 by Strapi will look like this

videochunk-What-Is-FRED-or-Federal-Reserve-Economic-Database-14

Using the DB ID seemed like a better alternative to using something as unreadable as UUIDs while still ensuring that the slug is unique. DB IDs can overlap between video/text chunks so some sort of identifier is needed - videochunk/textchunk as a prefix or affix (14v for video and 14t for text) seems like a reasonable option.